requirements.txt
click
gtfs-realtime-bindings
paho-mqtt
pyyaml
requests